third_party/blink/web_tests/fast/css/font-face-data-uri-expected.txt

Bug 49374 - Web Font specified with base64 data URI ending with EOT is not rendered. Each of the following pairs should look identical.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS document.getElementById('ref0').offsetWidth is document.getElementById('test0').offsetWidth
PASS document.getElementById('ref1').offsetWidth is document.getElementById('test1').offsetWidth
Proper data URI
Proper data URI

Data ends with EOT
Data ends with EOT
